Change: 日付差分表示をコンポーネントではなく関数に置き換え
